+++ src/vectorscope.c
@@ -278,12 +278,12 @@ static void create_graticule_vbuf(struct vss_source *src)
 	int stl_r = src->graticule_skintone_color & 0xFF;
 	switch(src->cm.colorspace) {
 		case 1: // BT.601
-			stl_u = RGB2U_601(stl_r, stl_g, stl_b);
-			stl_v = RGB2V_601(stl_r, stl_g, stl_b);
+			stl_u = (float)RGB2U_601(stl_r, stl_g, stl_b);
+			stl_v = (float)RGB2V_601(stl_r, stl_g, stl_b);
 			break;
 		default: // BT.709
-			stl_u = RGB2U_709(stl_r, stl_g, stl_b);
-			stl_v = RGB2V_709(stl_r, stl_g, stl_b);
+			stl_u = (float)RGB2U_709(stl_r, stl_g, stl_b);
+			stl_v = (float)RGB2V_709(stl_r, stl_g, stl_b);
 			break;
 	}
 	stl_norm = hypotf(stl_u-128.0f, stl_v-128.0f);
